pyproject.toml
[project]
name = "desktop-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "faster-whisper>=1.2.0",
    "mss>=10.1.0",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "pygetwindow>=0.0.9",
    "pynput>=1.8.1",
    "pytesseract>=0.3.13",
    "sounddevice>=0.5.3",
    "uiautomation>=2.0.29",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.16.0",
]
